Added another constructor, moved some includes/namespace
Moved include files to the .h just to keep them in once place.

Also added another constructor. This one uses an ostream object as a
parameter which allows us to send the output that should be displayed on
the LCD screen of the machine to something other than cout. The default
constructor initializes it to cout however. If you create a
VendingMachine object like this:
ofstream outFileTest("outTest.txt");

VendingMachine machine(outFileTest);

it will create a file that displays everything the VendingMachine class
itself used to output to the console.