Added Timer Class
Timer class objects store time of creation. Only one is needed for each
VendingMachine object as update time changes to current time.

We still need a checkTime function to call in the VendingMachine class
to make use of this however. This could be difficult as control will be
passed to the input function while time is passing and wait until a key
is entered. Unsure if this class can help in that case, but decided to
add just in case