Changed simulateMachine function, updated To Do list, added remove comment
To do list had listed that output should (in order I'm assuming) list
the simulated input and then the result printed to the display. This was
done by simply passing the object to both the VendingMachine and the
simulateMachine function and they alternate in writing to the file.

And for anyone who'd like to clean up the code a bit, I've been adding
"remove" to sections of code I'm not sure we need any longer. Mostly
things that seem to be leftovers from testing, though I commented out
the cout statement in simulateMachine. Uncommenting will make the
simulated keypresses print to the console as well as the
simOutputResults.txt