Extended Range of Buttons from 1-9 to 0-9
Consquences mean that the results from the input is not diferent. D1 and
A1 no longer have stock in them so the commands that recieved a drink
before now fail. The C0 and D0 tests now show the selection. Input file
needs remaking to make the stock changes more dramatic.